5月23号下午18时，增加咨询功能，功能完成，运行正常，服务器软件版本为开发板V523.1

+++ app/src/main/java/com/example/zbwx/model/Consult.java
@@ -7,7 +7,7 @@ import java.util.List;
 public class Consult {
     public int userID;
     public int adminID;
-    public long consultID; //数据库列头自增ID
+    public int consultID; //数据库列头自增ID
     public String title;
     public List<ConsultMsg> contentlist;//json数组字符串格式，内容为，datatime:xx type：xx  content:xx
